+++ frontend/src/app/components/cupom/cupom-form/cupom-form.component.ts
@@ -13,6 +13,7 @@ import { MatIconModule } from '@angular/material/icon';
 import { NgIf } from '@angular/common';
 import { Cupom } from '../../../models/cupom.model';
 import { MAT_DATE_LOCALE, provideNativeDateAdapter } from '@angular/material/core';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-cupom-form',
@@ -100,13 +101,32 @@ export class CupomFormComponent {
 
   excluir() {
     const cupom = this.formGroup.value;
-    this.cupomService.delete(cupom).subscribe({
-      next: () => {
-        this.router.navigateByUrl('/admin/cupons');
-      },
-      error: (e) => {
-        console.log('Erro ao excluir', JSON.stringify(e));
-      },
-    });
+    Swal.fire({
+          title: "Você tem certeza?",
+          text: "Vou não vai poder reverter isso!",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          confirmButtonText: "Sim, deletar!"
+        }).then((result) => {
+          if (result.isConfirmed) {
+            this.cupomService.delete(cupom).subscribe({
+              next: () => {
+                Swal.fire({
+                  title: "Deletado!",
+                  text: "Cupom deletado com sucesso!",
+                  icon: "success"
+                });
+                this.router.navigateByUrl('/', { skipLocationChange: true }).then(() => {
+                  this.router.navigate(['/admin/cupons']);
+                });
+              },
+              error: (e) => {
+                console.log('Erro ao excluir', JSON.stringify(e));
+              }
+            });
+          }
+        });
   }
 }
