+++ frontend/src/app/components/funcionario/funcionario-form/funcionario-form.component.ts
@@ -20,6 +20,7 @@ import { Telefone } from '../../../models/telefone.model';
 import { Cidade } from '../../../models/cidade.model';
 import { CidadeService } from '../../../services/cidade.service';
 import { Endereco } from '../../../models/endereco.model';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-funcionario-form',
@@ -171,13 +172,32 @@ export class FuncionarioFormComponent {
 
   excluir() {
     const funcionario = this.formGroup.value;
-    this.funcionarioService.delete(funcionario).subscribe({
-      next: () => {
-        this.router.navigateByUrl('/admin/funcionarios');
-      },
-      error: (e) => {
-        console.log('Erro ao excluir', JSON.stringify(e));
-      },
+    Swal.fire({
+      title: "Você tem certeza?",
+      text: "Vou não vai poder reverter isso!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Sim, deletar!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.funcionarioService.delete(funcionario).subscribe({
+          next: () => {
+            Swal.fire({
+              title: "Deletado!",
+              text: "Funcionario deletado com sucesso!",
+              icon: "success"
+            });
+            this.router.navigateByUrl('/', { skipLocationChange: true }).then(() => {
+              this.router.navigate(['/admin/funcionarios']);
+            });
+          },
+          error: (e) => {
+            console.log('Erro ao excluir', JSON.stringify(e));
+          }
+        });
+      }
     });
   }
